Extract `ReferenceFrame` into a separate unit
